+++ Assets/Mizore_Nekoyanagi/Util/MizoresPackageExporter/EditorScript/MizoresPackageExporterGUI/GUI_ExportPackage.cs
@@ -1,5 +1,5 @@
 ﻿using UnityEngine;
-using Const = MizoreNekoyanagi.PublishUtil.PackageExporter.MizoresPackageExporterConsts;
+using Const = MizoreNekoyanagi.PublishUtil.PackageExporter.ExporterConsts;
 using System.Collections.Generic;
 using System.Linq;
 using System.IO;
