+++ Assets/Mizore_Nekoyanagi/Util/MizoresPackageExporter/EditorScript/MizoresPackageExporterGUI/MizorePackageExporterEditor.cs
@@ -6,6 +6,7 @@ using UnityEditor;
 
 namespace MizoreNekoyanagi.PublishUtil.PackageExporter.ExporterEditor
 {
+
 #if UNITY_EDITOR
     [CustomEditor( typeof( MizoresPackageExporter ) ), CanEditMultipleObjects]
     public class MizoresPackageExporterEditor : Editor
@@ -16,33 +17,6 @@ namespace MizoreNekoyanagi.PublishUtil.PackageExporter.ExporterEditor
 
         private void OnEnable( ) {
             t = target as MizoresPackageExporter;
-        }
-        static string ToAssetsPath( string path ) {
-            string datapath = Application.dataPath;
-            if ( path.StartsWith( datapath ) ) {
-                path = path.Substring( datapath.Length - "Assets".Length );
-            }
-            return path;
-        }
-        public string BrowseButtons( string text ) {
-            string result = text;
-            if ( GUILayout.Button( ExporterTexts.TEXT_BUTTON_FOLDER, GUILayout.Width( 50 ) ) ) {
-                text = EditorUtility.OpenFolderPanel( null, t.ConvertDynamicPath( text ), null );
-                text = ToAssetsPath( text );
-                if ( string.IsNullOrEmpty( text ) == false ) {
-                    GUI.changed = true;
-                    result = text;
-                }
-            }
-            if ( GUILayout.Button( ExporterTexts.TEXT_BUTTON_FILE, GUILayout.Width( 50 ) ) ) {
-                text = EditorUtility.OpenFilePanel( null, t.ConvertDynamicPath( text ), null );
-                text = ToAssetsPath( text );
-                if ( string.IsNullOrEmpty( text ) == false ) {
-                    GUI.changed = true;
-                    result = text;
-                }
-            }
-            return result;
         }
         public override void OnInspectorGUI( ) {
             foreach ( var item in targets ) {
