+++ Assets/Mizore_Nekoyanagi/Util/MizoresPackageExporter/PackageExporter.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 837703c39835e26469e622d6cc8e8a99
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
