+++ Assets/Mizore_Nekoyanagi/Util/MizoresPackageExporter/PackageExporter/ExporterConsts.cs
@@ -1,0 +1,9 @@
+﻿namespace MizoreNekoyanagi.PublishUtil.PackageExporter
+{
+    public static class ExporterConsts
+    {
+        public const string ASSET_NAME = "MizoresPackageExporter";
+        public const string EXPORT_FOLDER_PATH = "MizorePackageExporter/";
+
+    }
+}
