+++ Assets/Mizore_Nekoyanagi/Util/MizoresPackageExporter/PackageExporter/ExporterConsts_Editor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5955806cdfd8f4d4ba728e34f50e701d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
