+++ Assets/Mizore_Nekoyanagi/Util/MizoresPackageExporter/PackageExporter/ExporterConsts_Keys.cs
@@ -1,0 +1,11 @@
+﻿namespace MizoreNekoyanagi.PublishUtil.PackageExporter
+{
+    public static class ExporterConsts_Keys
+    {
+        public const string KEY_NAME= "%name%";
+        public const string KEY_VERSION = "%version%";
+        public const string KEY_FORMATTED_VERSION = "%versionf%";
+        public const string KEY_PACKAGE_NAME= "%packagename%";
+        public const string KEY_BATCH_EXPORTER = "%batch%";
+    }
+}
