+++ Assets/Mizore_Nekoyanagi/Util/MizoresPackageExporter/PackageExporter/ExporterConsts_Keys.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7fc97f3d81e36fe40a45ddeea55be0df
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
